a) Added type encoding.
b) Added expression support.

c) Completed the box language interface

d) Completed the implementation of records.

e) Completed the implementation of the opaque handle structure.

f) As a result of c, d, e completed:
    1. the SNetOutxxx() functions
    2. the entity functions of the CRI (in entities.utc)
    3. the graph structure and its related functions

g) Added support for a uniform way of reporting errors and warnings.

h) Modified the graph index structure and functions so that the graph index
   structure does not "inherit" from the base type.

i) Modified the base type to be able to hold both places and contracts
   and to free the contracts automatically when destroyed. Modified other
   parts affected by this change accordingly.

j) Further developed the graph walker (file gw.utc) and related algorithms. As
   it is right now it is almost complete and operational. In particular:
        1. Walking functionality is complete
        2. It can deal with normal boxes, parallel, star, index split combinators,
           it can deal with multiplicity and flow-inheritance. Synchrocells
           are currently being ignored as the related functions are still incomplete.
           The same goes for filter boxes for which nothing has been implemented yet
           except that a graph-node has been partially defined (i.e. if a network
           has filter boxes everything will work ok but they will be ignored).
        3. Input/Output of records is complete.
        4. Index managemnt is complete (that is modify the index accordingly
           while "walking" as well as infimum index propagation).

k) Modified memory and place management so that errors in cases of failed
   allocations are handled within that module (so when functions such as
   SNetMemAlloc() are used there is no need to check whether they failed).
   Also added a very simple monitoring feature that keeps track of memory
   allocations and deallocations to help on detection of  memory leaks.

l) Modified the list structure so that it is able to hold items of types
   other than 'void*'. Due to this change the interface has slightly changed as well.

m) Added tests for the new features and modified test code to reflect all the above changes.

n) Modified the network with id 0 to allow for testing of the new features and other modifications.

o) Minor modifications to the build system mostly related to parameters for the compilers / tools used.

p) Fixed bugs and did some code formatting!!

As a general comment the state of the runtime in this revision is such that most of the required
functionality has been implemented!! It can run networks with the restriction that they do not
contain synchro-cells and filter boxes (they can but the runtime does nothing about them effectively
ignoring them).