Several changes / fixes / further development!
While implementing synchro-cells an issue was encountered regarding
flow-inheritance! In particular the S-Net documentation was specifying
that the 1st record inherits any excess fields. However the cons-list and
indexing schemes even though they can identify which records should sync
they cannot identify which one is the 1st in stream order! In order to
solve this several changes were performed on the cons-list and the
implementation of synchro-cells. These include making the list double-linked
and indexing the nodes (using the same index vector implementation as for
the graph) to identify their position in the cons-list. Having that info (meaning
being able to compare 2 cons-list nodes on their position on the cons-list) also
makes the implementation of synchro-cells simpler as no buffers for more than 1
record for each pattern is required!

After consulting with Clemens I found out fhat flow-inheritance on synchro-cells
was an issue under dispute and that the rule specified in the S-Net language
reference about it do not apply any more which makes the above to not be required.
However because it simplifies the implementation of synchro-cells even though there
is an issue with the "increasing length => required memory resources" of cons-node
indexes in cases where a "star" combinator is involved or in cases of multiplicity (or
even worse when both are combined) it will be left like this for now and further
investigation will be done later.

Records as a result of multiplicity are inserted "in-front" of the currently
processed node (for this the cons-list was required to be double-linked). This was
done to allow boxes with "infinite" multiplicity (not sure though if this is
required).

Finally several changes were done to the test application and in general
further developed several parts of the system!!