Added separate build system so that the runtime can be built for the duTC-PTL and build configurations were also created so that the runtime can be built for the MicroBlaze / Petalinux platform.
It needs to be different due to the extra types of files the
current implementation of duTC-PTL requires (e.g. "xdr" files) and the
tools used which might not be the same for other implementations or
with whatever (if ever) is standardized in the future.

Did appropriate modifications to the code so that it compiles also
with duTC-PTL. In fact the code for allocating / deallocating "places"
has been changed and it now uses a new API that both versions of
uTC-PTL (normal and distributed) provide for performing that operation
"portably" (at least with regards to the version of uTC-PTL used).

Added fixes / upgrades to the "normal" build system that the new
one for duTC-PTL had (since they follow the same design and have
common parts updates and fixes to one apply to the other as well).