Modified again the layout for the testing framework. In particular the different aspects of it (e.g. unit tests, snets, etc..) have been separated into multiple libraries. What previously consisted the "main" file of the test application is now the "runner" application for unit tests and test networks (and it links with the mentined libraries).
This separation was done because now there is a second application which
is a test SANE with which the related aspects of the runtime will be
tested later on (for now that application only contains dummy implementations
of the functions "required" by the SEP/SANE framework provided by duTC-PTL).
As with the "runner" this application will need to use functionality provided
by the mentioned libraries (thus it made sense to put that code into libraries).

Finally did some minor fixes that were causing warning for re-definitions of
pre-processor symbols in MicroBlaze builds.