CRI representation of boxes has changed: SNetBox now requires the name of the associated box function, which still is passed as function pointer, to be encoded as string as well. This new parameter is expected to be found as direct predecessor of the function pointer parameter, e.g.
SNetBox( instream, &boxfun, out_type) now becomes
SNetBox( instream, "boxfun", &boxfun, out_type)

The name of the boxfunction can either be set by meta-data element <boxfun>,
or, if this element is not set, by the name of the box as defined in the S-Net
source.

Codegeneration of the compiler has been updated accordingly.