Added the GW code ("gw/gw.utc" file) which was missing after the restructuring of the tree. This now has been modified to work with the latest changes introduced in the whole runtime! In fact most of the GW code has been refactored and the issues related to missing features (e.g. synchro-cells) and correct handling of flow-inheritance have been addressed.
Thus the GW is "functional" again and in terms of SNet features it does not
support filters! However because this code is still going through testing/debugging
it contains a lot of "debugging" code (e.g. debug prints) and needs some cleanup.

Did several modifications here and there as a result of adding (and modifying)
the GW code as well as bug fixes.

Some of the modification that took place were related to better intergration
with the SNet compiler. So now the compiler generated code does not require
any manual modifications and can be used as is. Thus keeping duplicate
copies of that (non-modified, modified) for the "test" networks is no longer
necessary so the "modified" copies have been removed (the build system does
not yet compile the ".snet" files so a copy of the generated ".c" file is
still required; this will be fixed in the future and these will also be
removed).