Fixed a GC bug in the collector task.
Bug: if the collector received a sync record on one of its inputs
with another source than the corresponding parallel dispatcher,
any incoming sort records on that stream would be treated wrongly.
The problem is that they are treated at all.
A sync record with an "external" source is now detected and forwarded
to the outgoing stream upon termination of the collector.

By the time the collector code is fairly complex and ugly and
would need some rewriting.