Fixed a flaw in cleanup of initialiser branches. This also affects GC;
Now, sort records are sent through the non-initialiser branches,
and at the collector, termination of the (static) collector
is done at the swapping of waitingset and readyset.
For GC triggered by the outtype of a sync-cell this is done as well
now, this led to simplification of the collector as sync records
that contain a flagged stream does not need to be stored -
the collector task code is unified again.